:recycle: : 유저 권한 설정 추가 in 그룹 삭제 API, 권한 체크 함수 리팩토링
