+++ src/main/java/com/capstone03/goldenglobe/CheckListAuthCheck.java
@@ -2,6 +2,8 @@ package com.capstone03.goldenglobe;
 
 import com.capstone03.goldenglobe.checkList.CheckList;
 import com.capstone03.goldenglobe.checkList.CheckListRepository;
+import com.capstone03.goldenglobe.listGroup.ListGroup;
+import com.capstone03.goldenglobe.listGroup.ListGroupRepository;
 import com.capstone03.goldenglobe.sharedList.SharedListRepository;
 import com.capstone03.goldenglobe.user.CustomUser;
 import lombok.RequiredArgsConstructor;
@@ -19,6 +21,7 @@ public class CheckListAuthCheck {
 
     private final CheckListRepository checkListRepository;
     private final SharedListRepository sharedListRepository;
+    private final ListGroupRepository listGroupRepository;
 
     public boolean hasAccessToCheckList(Long checkListId, Authentication auth) {
         // CheckList 존재 여부 확인
@@ -36,5 +39,18 @@ public class CheckListAuthCheck {
 
         // 체크리스트 공유 여부 확인
         return sharedListRepository.existsByList_ListIdAndUser_UserId(checkList.getListId(), authUserId);
+    }
+
+    public ListGroup findAndCheckAccessToGroup(Long groupId, Authentication auth) {
+        // 그룹 조회 및 존재하지 않을 경우 예외 처리
+        ListGroup listGroup = listGroupRepository.findByGroupId(groupId)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "일치하는 group_id가 없음"));
+
+        // 유저 권한 확인 절차
+        if (!hasAccessToCheckList(listGroup.getList().getListId(), auth)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "체크리스트에 접근할 수 없습니다.");
+        }
+
+        return listGroup;
     }
 }
